# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Optimization Service",
    "description": "Runs an agent against a TerminalBench subset in sandboxes, mines failures, and iteratively improves agent/agent.py via an LLM optimization loop.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Submit a benchmark optimization job",
        "operationId": "create_job_jobs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List jobs visible to the current user",
        "operationId": "list_jobs_jobs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobSummary"
                  },
                  "title": "Response List Jobs Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get job status and latest results",
        "operationId": "get_job_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/iterations": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List all iterations for a job",
        "operationId": "list_iterations_jobs__job_id__iterations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IterationSummary"
                  },
                  "title": "Response List Iterations Jobs  Job Id  Iterations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/iterations/{iteration_no}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get full detail for one iteration (traces, LLM artifacts)",
        "operationId": "get_iteration_detail_jobs__job_id__iterations__iteration_no__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "iteration_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Iteration No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IterationDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/agent-versions": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List all agent versions for a job",
        "operationId": "list_job_agent_versions_jobs__job_id__agent_versions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentVersionSummary"
                  },
                  "title": "Response List Job Agent Versions Jobs  Job Id  Agent Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/agent-versions/{version_no}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get one agent version (full agent.py + diff)",
        "operationId": "get_job_agent_version_jobs__job_id__agent_versions__version_no__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "version_no",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version No"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentVersionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel a queued or running job",
        "operationId": "cancel_job_jobs__job_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orgs": {
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Create an organization",
        "operationId": "create_org_orgs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orgs/{org_id}/users": {
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Create a user in an organization",
        "operationId": "create_user_orgs__org_id__users_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentVersionDetail": {
        "properties": {
          "version_no": {
            "type": "integer",
            "title": "Version No"
          },
          "parent_version_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Version No"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "created_by_iteration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Iteration",
            "description": "Iteration whose optimizer produced this version (None for v0)."
          },
          "diff": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diff",
            "description": "Unified diff against parent version."
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Full agent/agent.py source for this version."
          }
        },
        "type": "object",
        "required": [
          "version_no",
          "parent_version_no",
          "sha256",
          "content"
        ],
        "title": "AgentVersionDetail"
      },
      "AgentVersionSummary": {
        "properties": {
          "version_no": {
            "type": "integer",
            "title": "Version No"
          },
          "parent_version_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Version No"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "created_by_iteration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Iteration",
            "description": "Iteration whose optimizer produced this version (None for v0)."
          },
          "diff": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diff",
            "description": "Unified diff against parent version."
          }
        },
        "type": "object",
        "required": [
          "version_no",
          "parent_version_no",
          "sha256"
        ],
        "title": "AgentVersionSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobCreate": {
        "properties": {
          "task_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Ids",
            "description": "Benchmark task IDs to run. Defaults to the built-in subset.",
            "examples": [
              [
                "regex-log",
                "cobol-modernization"
              ]
            ]
          },
          "max_iterations": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Iterations",
            "default": 5
          },
          "patience": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Patience",
            "description": "Stop after this many consecutive iterations without score improvement.",
            "default": 3
          },
          "executor": {
            "type": "string",
            "enum": [
              "simulated",
              "harbor"
            ],
            "title": "Executor",
            "description": "simulated = deterministic fake benchmark (M1/M2/M4 dev); harbor = real Terminal-Bench via Harbor+Daytona (M3).",
            "default": "simulated"
          },
          "config": {
            "$ref": "#/components/schemas/JobExecutorConfig",
            "description": "Executor-specific overrides (agent_model, n_concurrent, etc.)."
          }
        },
        "type": "object",
        "title": "JobCreate"
      },
...
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Job ID (32-character lowercase hex)."
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "stop_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StopReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "task_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Task Ids"
          },
          "max_iterations": {
            "type": "integer",
            "title": "Max Iterations"
          },
          "patience": {
            "type": "integer",
            "title": "Patience"
          },
          "executor": {
            "type": "string",
            "title": "Executor"
          },
          "best_val_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Val Score"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "learnings": {
            "type": "string",
            "title": "Learnings"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "best_agent_version_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Agent Version No"
          },
          "iterations": {
            "items": {
              "$ref": "#/components/schemas/IterationSummary"
            },
            "type": "array",
            "title": "Iterations"
          },
          "latest_task_results": {
            "items": {
              "$ref": "#/components/schemas/TaskResultOut"
            },
            "type": "array",
            "title": "Latest Task Results",
            "description": "Structured M1 result from the most recent completed iteration: per-task pass/fail status and failure summaries. Empty while the job is still queued/running with no finished iteration."
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "stop_reason",
          "task_ids",
          "max_iterations",
          "patience",
          "executor",
          "best_val_score",
          "created_at",
          "started_at",
          "finished_at",
          "learnings",
          "error"
        ],
        "title": "JobResponse",
        "description": "Job status plus structured benchmark results when available.\n\nWhen `status` is `completed`, `latest_task_results` is the M1 deliverable:\nwhich tasks passed/failed and a summary of observed failures.",
        "examples": [
          {
            "best_agent_version_no": 1,
            "best_val_score": 0.5,
            "created_at": "2026-07-21T12:00:00Z",
            "executor": "simulated",
            "finished_at": "2026-07-21T12:01:00Z",
            "id": "a1b2c3d4e5f6789012345678abcdef01",
            "iterations": [],
            "latest_task_results": [
              {
                "reward": 1.0,
                "status": "passed",
                "task_id": "regex-log"
              },
              {
                "failure_summary": "Simulated failure on cobol-modernization: agent did not explore environment or verify solution.",
                "reward": 0.0,
                "status": "failed",
                "task_id": "cobol-modernization"
              }
            ],
            "learnings": "",
            "max_iterations": 2,
            "patience": 2,
            "started_at": "2026-07-21T12:00:01Z",
            "status": "completed",
            "stop_reason": "max_iterations",
            "task_ids": [
              "regex-log",
              "cobol-modernization"
            ]
          }
        ]
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus"
      },
      "JobSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Job ID (32-character lowercase hex)."
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "stop_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StopReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "task_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Task Ids"
          },
          "max_iterations": {
            "type": "integer",
            "title": "Max Iterations"
          },
          "patience": {
            "type": "integer",
            "title": "Patience"
          },
          "executor": {
            "type": "string",
            "title": "Executor"
          },
          "best_val_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Val Score"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "stop_reason",
          "task_ids",
          "max_iterations",
          "patience",
          "executor",
          "best_val_score",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "JobSummary"
      },
...
      "OrganizationCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "examples": [
              "acme-corp"
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "OrganizationCreate"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "JWT access token from POST /auth/login or /auth/refresh.",
        "scheme": "bearer"
      }
    }
  }
}